django/requirements.txt
asgiref
autopep8
django==2.2.10
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
django-import-export
beautifulsoup4
lxml
wheel
bs4
pandas
yfinance
psycopg2==2.8.6
psycopg2-binary==2.8.6
djangorestframework
django-cors-headers
